fs: export 2 functions in checkExist.js
