+++ src/fs/create.js
@@ -1,6 +1,7 @@
 import fs from 'fs/promises';
 import path from 'path';
 import {fileURLToPath} from 'url'
+import {checkFileExists} from "./checkExist.js"
 
 const url = fileURLToPath(import.meta.url)
 const file = path.join(path.dirname(url),'files','fresh.txt')
@@ -9,7 +10,7 @@ const create = async () => {
     try {
         // const fd = await fs.open(file);
         // throw new Error('FS operation failed')
-        const srcexist = await checkPathExists(file)
+        const srcexist = await checkFileExists(file)
         if (srcexist) {
             throw new Error('FS operation failed')
         }
@@ -22,8 +23,3 @@ const create = async () => {
 };
 
 await create();
-async function checkPathExists(path) {
-    return fs.stat(path)
-        .then((res) => res.isFile())
-        .catch((err) => err.code === 'ENOENT' ? false : err);
-}
