+++ src/fs/delete.js
@@ -1,13 +1,13 @@
 import fs from 'fs/promises';
 import path from 'path';
 import {fileURLToPath} from 'url'
+import {checkFileExists} from "./checkExist.js"
 
 const url = fileURLToPath(import.meta.url)
 const srcfile = path.join(path.dirname(url),'files','fileToRemove.txt')
 const remove = async () => {
-    const srcexist = await checkPathExists(srcfile)
     try {
-        const srcexist = await checkPathExists(srcfile)
+        const srcexist = await checkFileExists(srcfile)
         if (!srcexist) {
             throw new Error('FS operation failed')
         }
@@ -21,9 +21,3 @@ const remove = async () => {
 };
 
 await remove();
-
-async function checkPathExists(path) {
-    return fs.stat(path)
-        .then((res) => res.isFile())
-        .catch((err) => err.code === 'ENOENT' ? false : err);
-}
