+++ src/fs/list.js
@@ -1,6 +1,7 @@
 import fs from 'fs/promises';
 import path from 'path';
 import {fileURLToPath} from 'url'
+import {checkPathExists} from "./checkExist.js"
 
 const url = fileURLToPath(import.meta.url)
 const srcdir = path.join(path.dirname(url),'files')
@@ -22,8 +23,3 @@ const list = async () => {
 
 await list();
 
-function checkPathExists(path) {
-    return fs.stat(path)
-        .then((res) => res.isDirectory())
-        .catch((err) => err.code === 'ENOENT' ? false : err);
-}
